+++ ir_attachment_url/models/binary_fields.py
@@ -30,9 +30,9 @@ def get_mimetype_and_optional_content_by_url(url):
 
 class Binary(fields.Binary):
 
-    def write(self, records, value, create=False):
+    def write(self, records, value):
         domain = [
-            ('res_model', '=', records._name),
+            ('res_model', '=', self.model_name),
             ('res_field', '=', self.name),
             ('res_id', 'in', records.ids),
         ]
@@ -79,7 +79,7 @@ class Binary(fields.Binary):
                 else:
                     atts.unlink()
         else:
-            super(Binary, self).write(records, value, create=create)
+            super(Binary, self).write(records, value)
 
 
 fields.Binary = Binary