+++ ir_config_parameter_multi_company/models/__init__.py
@@ -1,3 +1,2 @@
 from . import ir_config_parameter
-from . import res_users
 from . import ir_property