+++ ir_config_parameter_multi_company/models/ir_config_parameter.py
@@ -1,11 +1,9 @@
 import logging
 
 from odoo import models, fields, api, tools
-from odoo.addons.base.ir.ir_config_parameter import IrConfigParameter as IrConfigParameterOriginal
+from odoo.addons.base.models.ir_config_parameter import IrConfigParameter as IrConfigParameterOriginal
 
 _logger = logging.getLogger(__name__)
-
-DATABASE_SECRET_KEY = 'database.secret'
 
 # params that has to be shared across all companies
 SHARED_KEYS = [