+++ ir_config_parameter_multi_company/models/ir_property.py
@@ -7,6 +7,6 @@ class IrProperty(models.Model):
     @api.multi
     def write(self, vals):
         res = super(IrProperty, self).write(vals)
-        field = self.env.ref('base.field_ir_config_parameter_value')
+        field = self.env.ref('base.field_ir_config_parameter__value')
         self._update_db_value_website_dependent(field)
         return res